% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landscape}
\alias{landscape}
\title{Landscape}
\format{
A matrix with 60 rows and 2 columns:
\describe{
  \item{x}{x coordinate}
  \item{y}{y coordiante}
  ...
}
}
\usage{
landscape
}
\description{
xy coordinates for 60 patches for the Mainland-Island landscape
}
\keyword{datasets}
