% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-foodweb.R
\name{load_foodweb}
\alias{load_foodweb}
\title{Load Food Web}
\usage{
load_foodweb(fw_file, sp_file)
}
\arguments{
\item{fw_file}{file path of the food web adjacency table.}

\item{sp_file}{file path of the species data.frame.}
}
\value{
an igraph object.
}
\description{
Load Food Web
}
\examples{
i <- 1
path <- file.path("data", "foodwebs")
sp_file <- file.path(path, paste0("BodyMass_", i, ".out"))
fw_file <- file.path(path, paste0("web_", i, ".out"))
fw <- load_foodweb(fw_file, sp_file)
}
