% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-lanscape.R
\name{load_landscape}
\alias{load_landscape}
\title{Load Landscape}
\usage{
load_landscape(ls_file)
}
\arguments{
\item{ls_file}{file path of the food web adjacency table.}
}
\value{
a matrix with xy coordinates.
}
\description{
Load Landscape
}
\examples{
\dontrun{
list.files("data/landscapes", full.names = TRUE)
i <- 1
path <- file.path("data", "landscapes")
ls_file <- file.path(path, paste0("RGG_", i, ".out"))
ls <- load_landscape(ls_file)
}
}
