% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_web.R
\name{load_timeseries}
\alias{load_timeseries}
\title{Load Time Series}
\usage{
load_timeseries(ts_path)
}
\arguments{
\item{ts_path}{path of the time series}
}
\description{
Load Time Series
}
\examples{
\dontrun{
f <- "~/Proj/remilio/data/output/remilio/6/timeseries_MI_1.out"
}
}
