% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_web.R
\name{local_webs}
\alias{local_webs}
\title{Local Community Composition}
\usage{
local_webs(timeseries, foodweb)
}
\arguments{
\item{timeseries}{data.frame created with load_ts().}

\item{foodweb}{igraph object created with load_foodweb().}
}
\value{
an igraph object.
}
\description{
Local Community Composition
}
