% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-metrics.R
\name{network}
\alias{network}
\title{Compute network metrics}
\usage{
network(g)
}
\arguments{
\item{g}{igraph object of the food web.}
}
\value{
a data.frame with network metrics.
}
\description{
Compute network metrics
}
