% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-mems.R
\name{select_mems}
\alias{select_mems}
\title{Select MEMs}
\usage{
select_mems(webs, landscape, output, d2 = 0.3, style = "B")
}
\arguments{
\item{webs}{list object obtained from local_web() function.}

\item{landscape}{a landscape, i.e. a matrix with xy coordinates.}

\item{output}{string, where to write output files. Only path and file name,
with no extension.}
}
\description{
Select MEMs
}
