% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeseries}
\alias{timeseries}
\title{Time series}
\format{
A data frame with 2400 rows and 4 features:
\describe{
  \item{patch}{patch ID}
  \item{species}{species ID}
  \item{time}{integration time step}
  \item{biomass}{biomass of species}
  ...
}
}
\usage{
timeseries
}
\description{
One time series of species biomasses.
}
\keyword{datasets}
