% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-metrics.R
\name{trophic_levels}
\alias{trophic_levels}
\title{Calculate trophic level}
\usage{
trophic_levels(g)
}
\arguments{
\item{g}{igraph object of the food web.}
}
\value{
numeric vector of trophic levels.
}
\description{
Calculate trophic level
}
